requirements.txt
pdfplumber==0.7.1
PyAutoGUI==0.9.53
pyttsx3==2.90
pywhatkit==5.3
wikipedia==1.4.0
transformers~=4.22.1
googlesearch-python
winshell~=0.6